Added the pretty brackets feature for "()" &"[]"
Whenever a bracket is inserted a corresponding invisible bracket is inserted at either ends of the expression or the current block. Seeking left and seeking right are updated accordingly making the cursor always stay in between the invisible brackets. Backspace deletion is also managed making the invisible bracket delete whenever it's pairing bracket is deleted.